carformer/requirements.txt
PyYAML
torch>=2.0.0
transformers>4.30
tokenizers
plotly
omegaconf==2.3.0
hydra-core==1.2.0
mpi4py==3.1.4
#deepspeed==0.10.1
